requirements.txt
Flask==2.0.1
Flask-Mail==0.9.1
python-dotenv==0.19.0
Werkzeug==2.0.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
click==8.0.1
gunicorn==20.1.0